% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_inputs}
\alias{check_inputs}
\title{Checks inputs to \code{codiv()}}
\usage{
check_inputs(
  Host_tree,
  Symbiont_tree,
  Host_to_Symbiont_df,
  min_hosts,
  min_symbiont_tips,
  span_fraction,
  permutations
)
}
\arguments{
\item{Host_tree}{A binary phylogenetic tree of hosts}

\item{Symbiont_tree}{A binary phylogenetic tree of symbionts}

\item{Host_to_Symbiont_df}{A \code{data.frame()} that pairs all of the tips in the \code{Symbiont_tree} with the tip in the \code{Host_tree} it was isolated from.}

\item{min_hosts}{The minimum number hosts for a node to be evaluated (must be greater than 3)}

\item{min_symbiont_tips}{The minimum number symbiont tips for a node to be evaluated (recommend 7 or more)}

\item{span_fraction}{The fraction of the \code{Symbiont_tree} included in the codiv scan}

\item{permutations}{The number of permutations (recommend at least 99)}
}
\value{
Tests to make sure the inputs are acceptable
}
\description{
Checks inputs to \code{codiv()}
}
\examples{
check_inputs(Host_tree, Symbiont_tree, Host_to_Symbiont_df, 3, 7, 0.25, 99)

}
