% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codiv.R
\name{codiv}
\alias{codiv}
\title{Co-Diversification Scan Between Hosts and Symbionts}
\usage{
codiv(
  Host_tree,
  Symbiont_tree,
  Host_to_Symbiont_df,
  min_hosts = 3,
  min_symbiont_tips = 7,
  span_fraction = 0.1,
  permutations = 99,
  seed = 8675309,
  verbose = TRUE,
  Save_fp,
  methods = c("hommola", "paco", "parafit"),
  focus_hosts,
  subtree_features = TRUE,
  cores = 4,
  continue = TRUE
)
}
\arguments{
\item{Host_tree}{}

\item{Symbiont_tree}{}

\item{Host_to_Symbiont_df}{}

\item{min_hosts}{}

\item{min_symbiont_tips}{}

\item{span_fraction}{}

\item{permutations}{}

\item{seed}{}

\item{verbose}{}

\item{Save_fp}{}

\item{methods}{}

\item{focus_hosts}{}

\item{subtree_features}{}

\item{cores}{}

\item{continue}{}
}
\description{
Co-Diversification Scan Between Hosts and Symbionts
}
