% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hommola.R
\name{hommola}
\alias{hommola}
\title{Calculates the correlation between host and symbiont trees}
\usage{
hommola(i_host_subtree_dist, i_symbiont_subtree_dist, i_host_to_symbiont_df)
}
\arguments{
\item{i_host_subtree_dist}{}

\item{i_symbiont_subtree_dist}{}

\item{i_host_to_symbiont_df}{}
}
\description{
Calculates the correlation between host and symbiont trees
}
