% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hommola_wf.R
\name{hommola_wf}
\alias{hommola_wf}
\title{Calculates the Hommola Correlation with permutations}
\usage{
hommola_wf(
  i_host_subtree_dist,
  i_symbiont_subtree_dist,
  i_host_to_symbiont_df,
  permutations,
  seed
)
}
\arguments{
\item{i_host_subtree_dist}{}

\item{i_symbiont_subtree_dist}{}

\item{i_host_to_symbiont_df}{}

\item{permutations}{The number of permutations (recommend at least 99)}

\item{seed}{A numeric, used to \code{set.seed()}}
}
\description{
Calculates the Hommola Correlation with permutations
}
