% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_symbiont_links.R
\name{host_symbiont_links}
\alias{host_symbiont_links}
\title{Changes the host-to-symbiont \code{data.frame()} to wide-format}
\usage{
host_symbiont_links(Host_to_Symbiont_df)
}
\arguments{
\item{Host_to_Symbiont_df}{A \code{data.frame()} that includes columns named
"Host" and "Symbiont"}
}
\value{
A wide-format host-to-symbiont \code{data.frame()}
}
\description{
Changes the host-to-symbiont \code{data.frame()} to wide-format
}
\examples{
host_symbiont_links(Host_to_Symbiont_df)

}
