% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paco_wf.R
\name{paco_wf}
\alias{paco_wf}
\title{Runs PACo}
\usage{
paco_wf(
  i_host_subtree,
  i_symbiont_subtree,
  i_host_to_symbiont_mat,
  permutations,
  seed
)
}
\arguments{
\item{i_host_subtree}{}

\item{i_symbiont_subtree}{}

\item{i_host_to_symbiont_mat}{}

\item{permutations}{}

\item{seed}{}
}
\description{
Runs PACo
}
